delete v2 chains in v1 and move code around
