+++ npm/pkg/dataplane/policies/chain-management_linux.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 
 	"github.com/Azure/azure-container-networking/npm/metrics"
-	"github.com/Azure/azure-container-networking/npm/pkg/dataplane/ioutil"
 	"github.com/Azure/azure-container-networking/npm/util"
 	npmerrors "github.com/Azure/azure-container-networking/npm/util/errors"
+	ioutil "github.com/Azure/azure-container-networking/npm/util/osutil_linux"
 	"k8s.io/klog"
 	utilexec "k8s.io/utils/exec"
 )
@@ -24,11 +24,6 @@ const (
 	couldntLoadTargetErrorCode int = 2 // Couldn't load target `AZURE-NPM-EGRESS':No such file or directory
 
 	minLineNumberStringLength int = 3 // TODO transferred from iptm.go and not sure why this length is important, but will update the function its used in later anyways
-
-	azureChainGrepPattern   string = "Chain AZURE-NPM"
-	minAzureChainNameLength int    = len("AZURE-NPM")
-	// the minimum number of sections when "Chain NAME (1 references)" is split on spaces (" ")
-	minSpacedSectionsForChainLine int = 2
 )
 
 var (
@@ -53,7 +48,6 @@ var (
 		util.IptablesNewState,
 	}
 
-	errInvalidGrepResult                      = errors.New("unexpectedly got no lines while grepping for current Azure chains")
 	deprecatedJumpFromForwardToAzureChainArgs = []string{
 		util.IptablesForwardChain,
 		util.IptablesJumpFlag,
@@ -170,7 +164,7 @@ func (pMgr *PolicyManager) bootup(_ []string) error {
 		}
 	}
 
-	currentChains, err := pMgr.allCurrentAzureChains()
+	currentChains, err := ioutil.AllCurrentAzureChains(pMgr.ioShim.Exec, defaultlockWaitTimeInSeconds)
 	if err != nil {
 		return npmerrors.SimpleErrorWrapper("failed to get current chains for bootup", err)
 	}
@@ -384,45 +378,6 @@ func (pMgr *PolicyManager) chainLineNumber(chain string) (int, error) {
 		return lineNum, nil
 	}
 	return 0, nil
-}
-
-func (pMgr *PolicyManager) allCurrentAzureChains() (map[string]struct{}, error) {
-	iptablesListCommand := pMgr.ioShim.Exec.Command(util.Iptables,
-		util.IptablesWaitFlag, defaultlockWaitTimeInSeconds, util.IptablesTableFlag, util.IptablesFilterTable,
-		util.IptablesNumericFlag, util.IptablesListFlag,
-	)
-	grepCommand := pMgr.ioShim.Exec.Command(ioutil.Grep, azureChainGrepPattern)
-	searchResults, gotMatches, err := ioutil.PipeCommandToGrep(iptablesListCommand, grepCommand)
-	if err != nil {
-		return nil, npmerrors.SimpleErrorWrapper("failed to get policy chain names", err)
-	}
-	if !gotMatches {
-		return nil, nil
-	}
-	lines := strings.Split(string(searchResults), "\n")
-	if len(lines) == 1 && lines[0] == "" {
-		// this should never happen: gotMatches is true, but there is no content in the searchResults
-		return nil, errInvalidGrepResult
-	}
-	lastIndex := len(lines) - 1
-	lastLine := lines[lastIndex]
-	if lastLine == "" {
-		// remove the last empty line (since each line ends with a newline)
-		lines = lines[:lastIndex] // this line doesn't impact the array that the slice references
-	} else {
-		klog.Errorf(`while grepping for current Azure chains, expected last line to end in "" but got [%s]. full grep output: [%s]`, lastLine, string(searchResults))
-	}
-	chainNames := make(map[string]struct{}, len(lines))
-	for _, line := range lines {
-		// line of the form "Chain NAME (1 references)"
-		spaceSeparatedLine := strings.Split(line, " ")
-		if len(spaceSeparatedLine) < minSpacedSectionsForChainLine || len(spaceSeparatedLine[1]) < minAzureChainNameLength {
-			klog.Errorf("while grepping for current Azure chains, got unexpected line [%s] for all current azure chains. full grep output: [%s]", line, string(searchResults))
-		} else {
-			chainNames[spaceSeparatedLine[1]] = struct{}{}
-		}
-	}
-	return chainNames, nil
 }
 
 func onMarkSpecs(mark string) []string {
