+++ npm/util/osutil_linux/grep_test.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 package ioutil
 
 import (
