+++ npm/util/osutil_linux/restore.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 package ioutil
 
 import (
